Refine numeric type generation in API documentation
Update the API SDK generator to distinguish between Go integer (int, int64,
uint64) and floating-point (float64) types when generating documentation
and type definitions. This improves clarity in the generated API.md and
ensures more accurate type mapping for generated SDKs.

+++ sdk/swift/Sources/CaicSDK/Types.swift
@@ -635,13 +635,13 @@ public struct EventRateLimit: Codable {
 public struct EventStats: Codable {
     public let ts: Int
     public let cpuPerc: Double
-    public let memUsed: uint64
-    public let memLimit: uint64
+    public let memUsed: Int
+    public let memLimit: Int
     public let memPerc: Double
-    public let netRx: uint64
-    public let netTx: uint64
-    public let blockRead: uint64
-    public let blockWrite: uint64
+    public let netRx: Int
+    public let netTx: Int
+    public let blockRead: Int
+    public let blockWrite: Int
     public let diskUsed: Int
 }
